+++ tests/test_basename_empty.c
@@ -1,0 +1,21 @@
+#include "fsio.h"
+#include "test.h"
+#include <stdlib.h>
+#include <string.h>
+
+
+void test_impl()
+{
+  char *path = fsio_basename("");
+
+  assert_true(!strcmp(path, ""));
+
+  free(path);
+}
+
+
+int main()
+{
+  test_run(test_impl);
+}
+