+++ tests/test_dirname_null.c
@@ -1,0 +1,17 @@
+#include "fsio.h"
+#include "test.h"
+
+
+void test_impl()
+{
+  char *path = fsio_dirname(NULL);
+
+  assert_true(path == NULL);
+}
+
+
+int main()
+{
+  test_run(test_impl);
+}
+