+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### v0.1.3
 
+* New dirname function.
+* New basename function.
 * New fsio_copy_file functions.
 * New fsio_path_exists function.
 * New fsio_file_exists function.